plugins/redshift/requirements.txt
aws_cdk.core~=1.67.0
aws-cdk.aws-s3~=1.67.0
aws-cdk-aws-lambda~=1.67.0
aws-cdk-aws-redshift~=1.67.0
